requirements.txt
psycopg2-binary==2.9.3
python_dotenv==0.19.2
dash==2.1.0
pandas==1.4.1
console-menu==0.7.0
